Class "WmlMonitoringPoint"

A monitoring point is the feature of interest defined for WaterML observations, i.e. a monitoring point represents the real world feature for which observations are taken. This may be, for example, the position of a stream flow sensor at a river.

Objects from the Class

Objects can be created by calls of the form new("WmlMonitoringPoint", ...).

Slots

id:

Object of class "character", the GML id of the point.

identifier:

Object of class "character", the identifier of the point.

names:

Object of class "character", the names of the point.

sampledFeatures:

Object of class "character", the identifiers of sampled features.

shape:

Object of class "SamsShape", the shape of the point.

timeZone:

Object of class "ANY", the time zone (if available).

verticalDatums:

Object of class "AN", the vertical datum (if available).

Extends

Class "GmlFeature", directly. Class "GmlFeatureOrNULL", by class "GmlFeature", distance 2. Class "GmlFeatureOrGmlFeaturePropertyOrNULL", by class "GmlFeature", distance 2.

Methods

sosName

signature(x = "WmlMonitoringPoint"): access the name of the point.

Examples

showClass("WmlMonitoringPoint")

#> Class "WmlMonitoringPoint" [package "sos4R"]
#> 
#> Slots:
#>                                                                       
#> Name:  sampledFeatures              id      identifier           names
#> Class:       character       character       character       character
#>                                                       
#> Name:            shape  verticalDatums        timeZone
#> Class:       SamsShape             ANY             ANY
#> 
#> Extends: 
#> Class "GmlFeature", directly
#> Class "GmlFeatureOrNULL", by class "GmlFeature", distance 2
#> Class "GmlFeatureOrGmlFeaturePropertyOrNULL", by class "GmlFeature", distance 2