Classes and Construction Functions for Elements of the OWS Namespace

These classes represent elements from the OGC Web Services Common Specification and the OGC Web Services Common Standard.

OwsCapabilities(version, updateSequence = NA, owsVersion = sosDefaultGetCapOwsVersion,
                identification = NULL, provider = NULL, operations = NULL,
                contents = NULL, languages = NULL)
OwsException(exceptionCode, exceptionText = c(), locator = as.character(NA))
OwsExceptionReport(version, lang = as.character(NA), exceptions = list(NA))
OwsGetCapabilities(service, acceptVersions, sections = sosDefaultGetCapSections,
                   acceptFormats = sosDefaultGetCapAcceptFormats,
                   updateSequence = c(as.character(NA)),
                   owsVersion = sosDefaultGetCapOwsVersion, acceptLanguages = c(NA))
OwsOperation(name, DCPs, parameters = list(NA), constraints = list(NA),
             metadata = list(NA))
OwsOperationsMetadata(operations, parameters = list(NA), constraints = list(NA),
                      extendedCapabilities = xmlNode(NA))
OwsRange(minimumValue = as.character(NA), maximumValue = as.character(NA),
         rangeClosure = as.character(NA), spacing = as.character(NA))
OwsServiceProvider(providerName, providerSite = as.character(NA),
                   serviceContact = xmlNode(NA))
OwsServiceIdentification(serviceType, serviceTypeVersion, profile = c(NA), title,
                         abstract = c(NA), keywords = c(NA), fees = as.character(NA),
                         accessConstraints = c(NA))

Arguments

abstract
Brief narrative description of this server, normally available for display to a human.

acceptFormats
Unordered character vector of zero or more response formats desired by client, with preferred formats listed first.

acceptLanguages
Unordered character vector of zero or more languages desired by client, with preferred formats listed first. Only OWS 2.0.0!

acceptVersions
Comma-separated prioritized sequence of one or more specification versions accepted by client, with preferred versions listed first.

accessConstraints
Access constraints that should be observed to assure the protection of privacy or intellectual property, and any other restrictions on retrieving or using data from or otherwise using a server.

constraints
Constraint on valid domain of a non-parameter quantity that applies to an operation which a server implements.

contents
The provider section of a capabilities document, object of class OwsContentsOrNULL.

DCPs
Information for a Distributed Computing Platform (DCP) supported for an operation.

exceptionCode
The code attribute of an OWS Exception, see OwsExceptionsData.

exceptions
The list of OwsException in a OwsExceptionReport.

exceptionText
The text element of an OWS Exception, see OwsExceptionsData.

extendedCapabilities
The possible contents of the ExtendedCapabilities subsection are not specified by the SOS standard.

fees
Fees and terms for using a server, including the monetary units as specified in ISO 4217.

identification
The identification section of a capabilities document, object of class OwsServiceIdentificationOrNULL.

keywords
Unordered list of one or more commonly used or formalised word(s) or phrase(s) used to describe a server.

lang
The code attribute of an OWS Exception.

languages
The languages section of a capabilities document, currently an object of class XMLAbstractNode.

locator
The locator attribute of an OWS Exception, see OwsExceptionsData.

maximumValue
Maximum value of a range (numeric parameter).

metadata
Metadata about an operation and its implementation.

minimumValue
Minimum value of a range (numeric parameter).

name
Name of an operation (request) (for example, GetCapabilities).

operations
A list of objects of class OwsOperation in a OperationsMetadata object. The provider section of a capabilities document.

owsVersion
The used OWS specification version.

parameters
Parameter valid domain that applies to an operation which a server implements.

profile
Identifier of OGC Web Service (OWS) Application Profile.

providerName
Unique identifier for service provider organization.

providerSite
Reference to the most relevant web site of a service provider.

provider
The provider section of a capabilities document, object of class OwsServiceProviderOrNULL.

rangeClosure
Specifies which of minimum and maximum values are included in this range; include when not default of “closed” range. Possible values are closed, open, open-closed, or closed-open.

sections
Unordered character vector of zero or more names of sections of service metadata document to be returned in service metadata document.

serviceContact
Information for contacting service provider.

service
Service type identifier text.

serviceType
A service type name from registry of services.

serviceTypeVersion
Version of a service type implemented by a server.

spacing
Regular distance or spacing between allowed values in this range; include when range is not continuous.

title
Title of a server, normally used for display to a human.

updateSequence
Service metadata document version, value is "increased" whenever any change is made in complete service metadata document. This can be used to request a certain version of a metadata document. Parameter is found in both request and reponse, but may not be supported by a service.

version
The version of the document.

Objects from the Class

Objects can be created by calling the construction functions, e.g. in the form OwsCapabilities_1.1.0(...), OwsContents(...) or OwsException(...).

The following classes are virtual and no objects may be created from it: OwsContentsOrNULL, OwsServiceIdentificationOrNULL, OwsServiceProviderOrNULL, OwsOperationsMetadataOrNULL.

Slots

Capabilities:

contents:

Object of class "OwsContentsOrNULL", the contents section of a capabilities document.

identification:

Object of class "OwsServiceIdentificationOrNULL", the identification section of a capabilities document.

languages:

Object of class "XMLAbstractNode", the languages section of a capabilities document, only OWS 2.0.0!

operations:

Object of class "OwsOperationsMetadataOrNULL", the operations section of capabilities document.

owsVersion:

Object of class "character", the used version of OWS.

provider:

Object of class "OwsServiceProviderOrNULL", the provider section of a capabilities document.

updateSequence:

Object of class "character", the updateSequence attribute of a capabilities document.

version:

Object of class "character", the version of the described service in a capabilities document.

exceptionCode:

Object of class "character", the code attribute of an OWS Exception, see OwsExceptionsData.

exceptionText:

Object of class "vector", the text of an OWS Exception, see OwsExceptionsData.

locator:

Object of class "character", the locator attribute of an OWS Exception, see OwsExceptionsData.

version:

Object of class "character", the version of an OWS Exception, see OwsExceptionsData.

lang:

Object of class "character", the code attribute of an OWS Exception, see OwsExceptionsData.

exceptions:

Object of class "list", the list of OwsException in a OwsExceptionReport.

sections:

Object of class "vector", unordered character vector of zero or more names of sections of service metadata document to be returned in service metadata document.

acceptFormats:

Object of class "vector", unordered character vector of zero or more response formats desired by client, with preferred formats listed first.

updateSequence:

Object of class "vector", service metadata document version.

acceptVersions:

Object of class "character", comma-separated prioritized sequence of one or more specification versions accepted by client, with preferred versions listed first.

service:

Object of class "character", the name of the service.

request:

Object of class "character", the name of the operation/request.

acceptLanguages:

Object of class "vector", an unordered character vector of zero or more languages desired by client, with preferred formats listed first. Only OWS 2.0.0!

name:

Object of class "character", name of an operation (request) (for example, GetCapabilities).

DCPs:

Object of class "list", information for a Distributed Computing Platform (DCP) supported for an operation.

parameters:

Object of class "list", parameter valid domain that applies to an operation which a server implements.

constraints:

Object of class "list", Constraint on valid domain of a non-parameter quantity that applies to an operation which a server implements.

metadata:

Object of class "list", metadata about an operation and its implementation.

operations:

Object of class "list", a list of objects of class OwsOperation in a OperationsMetadata object.

extendedCapabilities:

Object of class "XMLAbstractNode", the possible contents of the ExtendedCapabilities subsection.

minimumValue:

Object of class "character", minimum value of a range (numeric parameter).

maximumValue:

Object of class "character", maximum value of a range (numeric parameter).

rangeClosure:

Object of class "character", specifies which of minimum and maximum values are included in this range.

spacing:

Object of class "character", regular distance or spacing between allowed values in this range; included when range is not continuous.

serviceType:

Object of class "character", the service type name from registry of services.

serviceTypeVersion:

Object of class "vector", version of a service type implemented by the server.

profile:

Object of class "vector", identifier of OGC Web Service (OWS) Application Profile.

title:

Object of class "vector", title of the server, normally used for display to a human.

abstract:

Object of class "vector", brief narrative description of this server, normally available for display to a human.

keywords:

Object of class "vector", unordered list of one or more commonly used or formalised word(s) or phrase(s) used to describe a server.

fees:

Object of class "character", fees and terms for using a server, including the monetary units as specified in ISO 4217.

accessConstraints:

Object of class "vector", access constraints that should be observed to assure the protection of privacy or intellectual property, and any other restrictions on retrieving or using data from or otherwise using a server.

providerName:

Object of class "character", unique identifier for service provider organization.

providerSite:

Object of class "character", reference to the most relevant web site of a service provider.

serviceContact:

Object of class "XMLAbstractNode", information for contacting service provider.

Extends

OwsCapabilities_1.1.0: Class "OwsCapabilities", directly.

OwsCapabilities_2.0.0: Class "OwsCapabilities_1.1.0", directly. Class "OwsCapabilities", by class "OwsCapabilities_1.1.0", distance 2.

OwsGetCapabilities_1.1.0 Class "OwsGetCapabilities", directly. Class "OwsServiceOperation", by class "OwsGetCapabilities", distance 2.

OwsGetCapabilities_2.0.0: Class "OwsGetCapabilities_1.1.0", directly. Class "OwsGetCapabilities", by class "OwsGetCapabilities_1.1.0", distance 2. Class "OwsServiceOperation", by class "OwsGetCapabilities_1.1.0", distance 3.

OwsServiceIdentification: Class "OwsServiceIdentificationOrNULL", directly.

OwsServiceProvider: Class "OwsServiceProviderOrNULL", directly.

Methods

show

signature(object = "<NAME OF CLASS>"): Shows a human readable version of the object.

checkRequest

signature(service = "SOS", operation = "OwsGetCapabilities_1.1.0", verbose = "logical") or signature(service = "SOS", operation = "OwsGetCapabilities_2.0.0", verbose = "logical"): Check the given operation for validity and for compliance with the metadata of the given SOS.

encodeRequestKVP

signature(obj = "OwsGetCapabilities"): See link{encodeRequestKVP}.

encodeRequestSOAP

signature(obj = "OwsGetCapabilities"): See link{encodeRequestSOAP}.

encodeRequestXML

signature(obj = "OwsGetCapabilities"): See link{encodeRequestXML}.

encodeRequestKVP

signature(obj = "OwsGetCapabilities_2.0.0"): See link{encodeRequestKVP}.

encodeRequestKVP

signature(obj = "OwsGetCapabilities_1.1.0"): See link{encodeRequestKVP}.

sosRequest

signature(sos = "SOS", request = "OwsServiceOperation", verbose = "logical", inspect = "logical"): Send the given operation as a request to the given SOS.

Details

OwsServiceOperation is the top class which is eventually put into the request method, sosRequest(...).

Classes ending in ...OrNULL are used to model optional slots.

References

Whiteside A. (Ed.), OGC Web Services Common Specification, Open Geospatial Consortium Inc., OGC 06-121r3, Version: 1.1.0 with Corrigendum 1

Whiteside A., Greenwood, J. (Eds.), OGC Web Services Common Standard, Open Geospatial Consortium Inc., OGC 06-121r9, Version: 2.0.0

Examples

showClass("OwsCapabilities_1.1.0")

#> Class "OwsCapabilities_1.1.0" [package "sos4R"]
#> 
#> Slots:
#>                                                                     
#> Name:                  identification                       provider
#> Class: OwsServiceIdentificationOrNULL       OwsServiceProviderOrNULL
#>                                                                     
#> Name:                      operations                       contents
#> Class:    OwsOperationsMetadataOrNULL              OwsContentsOrNULL
#>                                                                     
#> Name:                         version                 updateSequence
#> Class:                      character                      character
#>                                      
#> Name:                      owsVersion
#> Class:                      character
#> 
#> Extends: "OwsCapabilities"
#> 
#> Known Subclasses: "OwsCapabilities_2.0.0", "SosCapabilities_1.0.0", "SosCapabilities_2.0.0"

showClass("OwsCapabilities_2.0.0")

#> Class "OwsCapabilities_2.0.0" [package "sos4R"]
#> 
#> Slots:
#>                                                                     
#> Name:                       languages                 identification
#> Class:                XMLAbstractNode OwsServiceIdentificationOrNULL
#>                                                                     
#> Name:                        provider                     operations
#> Class:       OwsServiceProviderOrNULL    OwsOperationsMetadataOrNULL
#>                                                                     
#> Name:                        contents                        version
#> Class:              OwsContentsOrNULL                      character
#>                                                                     
#> Name:                  updateSequence                     owsVersion
#> Class:                      character                      character
#> 
#> Extends: 
#> Class "OwsCapabilities_1.1.0", directly
#> Class "OwsCapabilities", by class "OwsCapabilities_1.1.0", distance 2

showClass("OwsCapabilities")

#> Class "OwsCapabilities" [package "sos4R"]
#> 
#> Slots:
#>                                                    
#> Name:         version updateSequence     owsVersion
#> Class:      character      character      character
#> 
#> Known Subclasses: 
#> Class "OwsCapabilities_1.1.0", directly
#> Class "OwsCapabilities_2.0.0", by class "OwsCapabilities_1.1.0", distance 2
#> Class "SosCapabilities_1.0.0", by class "OwsCapabilities_1.1.0", distance 2
#> Class "SosCapabilities_2.0.0", by class "OwsCapabilities_1.1.0", distance 2

showClass("OwsContents")

#> Class "OwsContents" [package "sos4R"]
#> 
#> Slots:
#>                       
#> Name:              xml
#> Class: XMLAbstractNode
#> 
#> Extends: "OwsContentsOrNULL"
#> 
#> Known Subclasses: "SosContents"

showClass("OwsContentsOrNULL")

#> Virtual Class "OwsContentsOrNULL" [package "sos4R"]
#> 
#> No Slots, prototype of class "NULL"
#> 
#> Known Subclasses: 
#> Class "NULL", directly
#> Class "OwsContents", directly
#> Class ".NULL", by class "NULL", distance 2, with explicit coerce

showClass("OwsException")

#> Class "OwsException" [package "sos4R"]
#> 
#> Slots:
#>                                                 
#> Name:  exceptionCode exceptionText       locator
#> Class:     character        vector     character

showClass("OwsExceptionReport")

#> Class "OwsExceptionReport" [package "sos4R"]
#> 
#> Slots:
#>                                        
#> Name:     version       lang exceptions
#> Class:  character  character       list

showClass("OwsGetCapabilities_1.1.0")

#> Class "OwsGetCapabilities_1.1.0" [package "sos4R"]
#> 
#> Slots:
#>                                                                   
#> Name:        sections  acceptFormats updateSequence acceptVersions
#> Class:         vector         vector         vector      character
#>                                                                   
#> Name:      owsVersion        service        request        version
#> Class:      character      character      character      character
#> 
#> Extends: 
#> Class "OwsGetCapabilities", directly
#> Class "OwsServiceOperation", by class "OwsGetCapabilities", distance 2
#> 
#> Known Subclasses: "OwsGetCapabilities_2.0.0"

showClass("OwsGetCapabilities_2.0.0")

#> Class "OwsGetCapabilities_2.0.0" [package "sos4R"]
#> 
#> Slots:
#>                                                                       
#> Name:  acceptLanguages        sections   acceptFormats  updateSequence
#> Class:          vector          vector          vector          vector
#>                                                                       
#> Name:   acceptVersions      owsVersion         service         request
#> Class:       character       character       character       character
#>                       
#> Name:          version
#> Class:       character
#> 
#> Extends: 
#> Class "OwsGetCapabilities_1.1.0", directly
#> Class "OwsGetCapabilities", by class "OwsGetCapabilities_1.1.0", distance 2
#> Class "OwsServiceOperation", by class "OwsGetCapabilities_1.1.0", distance 3

showClass("OwsGetCapabilities")

#> Class "OwsGetCapabilities" [package "sos4R"]
#> 
#> Slots:
#>                                                                   
#> Name:  acceptVersions     owsVersion        service        request
#> Class:      character      character      character      character
#>                      
#> Name:         version
#> Class:      character
#> 
#> Extends: "OwsServiceOperation"
#> 
#> Known Subclasses: 
#> Class "OwsGetCapabilities_1.1.0", directly
#> Class "OwsGetCapabilities_2.0.0", by class "OwsGetCapabilities_1.1.0", distance 2

showClass("OwsOperation")

#> Class "OwsOperation" [package "sos4R"]
#> 
#> Slots:
#>                                                                   
#> Name:         name        DCPs  parameters constraints    metadata
#> Class:   character        list        list        list        list

showClass("OwsOperationsMetadata")

#> Class "OwsOperationsMetadata" [package "sos4R"]
#> 
#> Slots:
#>                                                                      
#> Name:            operations           parameters          constraints
#> Class:                 list                 list                 list
#>                            
#> Name:  extendedCapabilities
#> Class:      XMLAbstractNode
#> 
#> Extends: "OwsOperationsMetadataOrNULL"

showClass("OwsRange")

#> Class "OwsRange" [package "sos4R"]
#> 
#> Slots:
#>                                                           
#> Name:  minimumValue maximumValue rangeClosure      spacing
#> Class:    character    character    character    character

showClass("OwsServiceIdentification")

#> Class "OwsServiceIdentification" [package "sos4R"]
#> 
#> Slots:
#>                                                                
#> Name:         serviceType serviceTypeVersion            profile
#> Class:          character             vector             vector
#>                                                                
#> Name:               title           abstract           keywords
#> Class:             vector             vector             vector
#>                                             
#> Name:                fees  accessConstraints
#> Class:          character             vector
#> 
#> Extends: "OwsServiceIdentificationOrNULL"

showClass("OwsServiceIdentificationOrNULL")

#> Virtual Class "OwsServiceIdentificationOrNULL" [package "sos4R"]
#> 
#> No Slots, prototype of class "NULL"
#> 
#> Known Subclasses: 
#> Class "NULL", directly
#> Class "OwsServiceIdentification", directly
#> Class ".NULL", by class "NULL", distance 2, with explicit coerce

showClass("OwsServiceOperation")

#> Virtual Class "OwsServiceOperation" [package "sos4R"]
#> 
#> Slots:
#>                                     
#> Name:    service   request   version
#> Class: character character character
#> 
#> Known Subclasses: 
#> Class "OwsGetCapabilities", directly
#> Class "SosDescribeSensor", directly
#> Class "SosGetObservation", directly
#> Class "SosGetObservationById", directly
#> Class "SosGetObservation_2.0.0", directly
#> Class "SosGetFeatureOfInterest_2.0.0", directly
#> Class "OwsGetCapabilities_1.1.0", by class "OwsGetCapabilities", distance 2
#> Class "OwsGetCapabilities_2.0.0", by class "OwsGetCapabilities", distance 3

showClass("OwsServiceProvider")

#> Class "OwsServiceProvider" [package "sos4R"]
#> 
#> Slots:
#>                                                       
#> Name:     providerName    providerSite  serviceContact
#> Class:       character       character XMLAbstractNode
#> 
#> Extends: "OwsServiceProviderOrNULL"

showClass("OwsServiceProviderOrNULL")

#> Virtual Class "OwsServiceProviderOrNULL" [package "sos4R"]
#> 
#> No Slots, prototype of class "NULL"
#> 
#> Known Subclasses: 
#> Class "NULL", directly
#> Class "OwsServiceProvider", directly
#> Class ".NULL", by class "NULL", distance 2, with explicit coerce

# TBD examples for construction functions