Class and Construction Function for om:Measurement Elements

Classes and construction functions for objects from the OGC Observations and Measurements specification.

OmMeasurement(samplingTime, procedure, observedProperty, featureOfInterest, result,
              metadata = NA, resultTime = NULL, resultQuality = NA,     parameter = NA)

Arguments

samplingTime
See slot description.

procedure
See slot description.

observedProperty
See slot description.

featureOfInterest
See slot description.

result
See slot description.

metadata
See slot description.

resultTime
See slot description.

resultQuality
See slot description.

parameter
See slot description.

Objects from the Class

Objects can be created by calls to the construction function of the form OmMeasurement(...).

Slots

featureOfInterest:

Object of class "GmlFeature": A feature of any type (ISO 19109, ISO 19101), which is a representation of the observation target, being the real-world object regarding which the observation is made.

metadata:

Object of class "ANY": Observation metadata.

observedProperty:

Object of class "SwePhenomenonProperty": Identifies or describes the phenomenon for which the observation result provides an estimate of its value. It must be a property associated with the type of the feature of interest.

parameter:

Object of class "ANY": A general event-specific parameter. This will typically be used to record environmental parameters, or event-specific sampling parameters that are not tightly bound to either the feature-of-interest or the procedure.

procedure:

Object of class "ANY": The description of a process used to generate the result. It must be suitable for the observed property.

result:

Object of class "ANY": Contains the value generated by the procedure. The type of the observation result must be consistent with the observed property, and the scale or scope for the value must be consistent with the quantity or category type.

resultQuality:

Object of class "ANY": Event specific quality of a result.

resultTime:

Object of class "GmlTimeObjectOrNULL": The time when the procedure associated with the observation act was applied. For some observations these are identical, in which case the resultTime may be omitted. However, there are important cases where they differ.

samplingTime:

Object of class "GmlTimeObjectOrNULL": The time that the result applies to the feature-of-interest. This is the time usually required for geospatial analysis of the result.

Extends

OmMeasurement: Class "OmObservation", directly. Class "OmObservationOrNULL", by class "OmObservation", distance 2.

Methods

sosResult

signature(obj = "OmMeasurement"): Get the data in the measurement as a data.frame.

sosGetCRS

signature(obj = "OmMeasurement"): Get the coordinate reference system used in the feature of interest.

Details

A Measurement contains a GmlMeasure.

Value

The construction functions return an object of the respective class.

References

Cox, S. (Ed.), Observations and Measurements - Part 1 - Observation schema, Open Geospatial Consortium Inc., OGC 07-022r1, Version: 1.0

See also

Examples

showClass("OmMeasurement")

#> Class "OmMeasurement" [package "sos4R"]
#> 
#> Slots:
#>                                            
#> Name:                                result
#> Class:                           GmlMeasure
#>                                            
#> Name:                          samplingTime
#> Class:                  GmlTimeObjectOrNULL
#>                                            
#> Name:                             procedure
#> Class:                                  ANY
#>                                            
#> Name:                      observedProperty
#> Class:          SwePhenomenonPropertyOrNULL
#>                                            
#> Name:                     featureOfInterest
#> Class: GmlFeatureOrGmlFeaturePropertyOrNULL
#>                                            
#> Name:                              metadata
#> Class:                                  ANY
#>                                            
#> Name:                            resultTime
#> Class:                  GmlTimeObjectOrNULL
#>                                            
#> Name:                         resultQuality
#> Class:                                  ANY
#>                                            
#> Name:                             parameter
#> Class:                                  ANY
#> 
#> Extends: 
#> Class "OmObservation", directly
#> Class "OmObservationOrNULL", by class "OmObservation", distance 2

# TBD examples for construction function

# TBD examples for sosResult