Class "OmOM_Observation"

Classes and construction functions for objects from the OGC Observations and Measurements specification, version 2.0.

Objects from the Class

Objects can be created by calls of the form new("OmOM_Observation", ...).

Slots

phenomenonTime:

Object of class "GmlTimeObjectOrNULL": The time that the result applies to the feature-of-interest. This is the time usually required for geospatial analysis of the result. Note: The phenomenonTime has been renamed to samplingTime in the O&M specification version 2.0. However, in order to be consistent with version 1.0 in SOS4R, we decided to leave the slot as is and just changed the parsing/encoding.

procedure:

Object of class "ANY": The description of a process used to generate the result. It must be suitable for the observed property.

observedProperty:

Object of class "SwePhenomenonProperty": Identifies or describes the phenomenon for which the observation result provides an estimate of its value. It must be a property associated with the type of the feature of interest.

featureOfInterest:

Object of class "GmlFeature": A feature of any type (ISO 19109, ISO 19101), which is a representation of the observation target, being the real-world object regarding which the observation is made.

result:

Object of class "ANY": Contains the value generated by the procedure. The type of the observation result must be consistent with the observed property, and the scale or scope for the value must be consistent with the quantity or category type.

metadata:

Object of class "ANY": Observation metadata.

resultTime:

Object of class "GmlTimeObjectOrNULL": The time when the procedure associated with the observation act was applied. For some observations these are identical, in which case the resultTime may be omitted. However, there are important cases where they differ.

resultQuality:

Object of class "ANY": Event specific quality of a result.

parameter:

Object of class "ANY": A general event-specific parameter. This will typically be used to record environmental parameters, or event-specific sampling parameters that are not tightly bound to either the feature-of-interest or the procedure.

Extends

Class "OmOM_ObservationOrNULL", directly.

Methods

sosCoordinates

signature(obj = "OmOM_Observation"): Accessor to the spatial cooridinates of an observation.

sosFeatureIds

signature(obj = "OmOM_Observation"): Accessor to the feature identifiers.

sosFeaturesOfInterest

signature(obj = "OmOM_Observation"): Accessor to the feature objects.

sosObservedProperties

signature(obj = "OmOM_Observation"): Accessor to the observed properties in the observation.

sosProcedures

signature(obj = "OmOM_Observation"): Accessor to the procedures in the observation.

sosResult

signature(obj = "OmOM_Observation"): Extract the data from the observation as a data.frame.

sosTime

signature(obj = "OmOM_Observation"): Accessor to the time period of the observation.

Examples

showClass("OmOM_Observation")

#> Class "OmOM_Observation" [package "sos4R"]
#> 
#> Slots:
#>                                            
#> Name:                        phenomenonTime
#> Class:                  GmlTimeObjectOrNULL
#>                                            
#> Name:                             procedure
#> Class:                                  ANY
#>                                            
#> Name:                      observedProperty
#> Class:          SwePhenomenonPropertyOrNULL
#>                                            
#> Name:                     featureOfInterest
#> Class: GmlFeatureOrGmlFeaturePropertyOrNULL
#>                                            
#> Name:                                result
#> Class:                                  ANY
#>                                            
#> Name:                              metadata
#> Class:                                  ANY
#>                                            
#> Name:                            resultTime
#> Class:                  GmlTimeObjectOrNULL
#>                                            
#> Name:                         resultQuality
#> Class:                                  ANY
#>                                            
#> Name:                             parameter
#> Class:                                  ANY
#> 
#> Extends: "OmOM_ObservationOrNULL"