Classes of the Namespace sa

Classes and construction functions for elements from the OGC specification “Observations and Measurements - Part 2 - Sampling Features”.

SaSamplingPoint(sampledFeatures, position, relatedObservation = list(NA),
                relatedSamplingFeature = list(NA), surveyDetails = NA, id = NA_character_)

Arguments

sampledFeatures
~~

position
~~

relatedObservation
~~

relatedSamplingFeature
~~

surveyDetails
~~

id
~~

Objects from the Class

Objects can be created by calls to the construction functions of the form SaSamplingPoint(...).

Slots

sampledFeatures:

Object of class "list" which contains the sampled features.

position:

Object of class "GmlPointProperty" which contains the position of a feature.

relatedObservation:

Object of class "list" which contains identifiers of related observations.

relatedSamplingFeature:

Object of class "list" which contains identifiers of related sampling features.

surveyDetails:

Object of class "ANY" which can contain information about survey details (which are currently no modeled in an R class).

id:

The identifier of a sampling elements (object of class "character").

shape:

Object of class "ANY" which could be used to model the shape of a sampling surface.

area:

Object of class "ANY" which could be used to model the area of a sampling surface.

Extends

SaSamplingPoint and SaSamplingSurface: Class "GmlFeature", directly. Class "GmlFeatureOrNULL", by class "GmlFeature", distance 2.

Methods

show

signature(object = "SaSamplingPoint"): ...

show

signature(object = "SaSamplingSurface"): ...

sosCoordinates

signature(object = "SaSamplingPoint"): access to the coordinates of the feature.

sosFeatureIds

signature(object = "SaSamplingPoint"): access to the identifier of the feature.

sosFeaturesOfInterest

signature(object = "SaSamplingPoint"): access to the feature of interest, i.e. the point itself.

Value

Construction functions: An object of the respective class.

References

Simon Cox (Ed.), Observations and Measurements - Part 2 - Sampling Features, OGC 07-002r3

Examples

showClass("SaSamplingPoint")

#> Class "SaSamplingPoint" [package "sos4R"]
#> 
#> Slots:
#>                                                                            
#> Name:         sampledFeatures               position     relatedObservation
#> Class:                   list       GmlPointProperty                   list
#>                                                                            
#> Name:  relatedSamplingFeature          surveyDetails                     id
#> Class:                   list                    ANY              character
#> 
#> Extends: 
#> Class "GmlFeature", directly
#> Class "GmlFeatureOrNULL", by class "GmlFeature", distance 2
#> Class "GmlFeatureOrGmlFeaturePropertyOrNULL", by class "GmlFeature", distance 2

# create sampling point
SaSamplingPoint(sampledFeatures = list("feature1", "feature2"),
                position = GmlPointProperty(href = "http://link.to/point"))

#> Object of class SaSamplingPoint; id: NA; position: Object of class GmlPointProperty; href: http://link.to/point; point: , relatedObservation: NA, relatedSamplingFeature: , surveyDetails: NA;
#> 	sampledFeatures: feature1, feature2