Classes and creation functions for Sampling Features

Sampling Feature classes.

SamsShape represents the geometry of a spatial sampling feature, that can be used as a feature of interest. Currently, only points are supported.

Objects from the Class

Objects can be created by calls to the creation functions:

SamsSamplingFeature(...)

SamsShape(...)

Slots

identifier:

Object of class "character" ~~

name:

Object of class "character" ~~

type:

Object of class "character" ~~

sampledFeature:

Object of class "character" ~~

shape:

Object of class "SamsShape" ~~

id:

Object of class "character" ~~

point:

Object of class "GmlPoint" ~~

id:

Object of class "character" ~~

srsName:

Object of class "character" ~~

srsDimension:

Object of class "integer" ~~

axisLabels:

Object of class "character" ~~

uomLabels:

Object of class "character" ~~

Extends

SamsShape extends "GmlGeometry", directly.

SamsSamplingFeature extends "GmlFeature", directly. SamsSamplingFeature extends "GmlFeatureOrNULL", by class "GmlFeature", distance 2. SamsSamplingFeature extends "GmlFeatureOrGmlFeaturePropertyOrNULL", by class "GmlFeature", distance 2.

Methods

sosCoordinates

signature(obj = "SamsSamplingFeature"): access to the coordinates of the feature.

sosFeatureIds

signature(obj = "SamsSamplingFeature"): access to the identifier of the feature.

sosFeaturesOfInterest

signature(obj = "SamsSamplingFeature"): access the feature itself.

sosName

signature(obj = "SamsSamplingFeature"): access a human-readable name of the feature.

References

https://www.opengeospatial.org/standards/om

Note

Schema: http://schemas.opengis.net/sampling/2.0/samplingFeature.xsd

Examples

showClass("SamsSamplingFeature")

#> Class "SamsSamplingFeature" [package "sos4R"]
#> 
#> Slots:
#>                                                                   
#> Name:      identifier           name           type sampledFeature
#> Class:      character      character      character      character
#>                                     
#> Name:           shape             id
#> Class:      SamsShape      character
#> 
#> Extends: 
#> Class "GmlFeature", directly
#> Class "GmlFeatureOrNULL", by class "GmlFeature", distance 2
#> Class "GmlFeatureOrGmlFeaturePropertyOrNULL", by class "GmlFeature", distance 2

showClass("SamsShape")

#> Class "SamsShape" [package "sos4R"]
#> 
#> Slots:
#>                                                                        
#> Name:         point           id      srsName srsDimension   axisLabels
#> Class:     GmlPoint    character    character      integer    character
#>                    
#> Name:     uomLabels
#> Class:    character
#> 
#> Extends: "GmlGeometry", "SamsShapeOrNULL"