Classes and Construction Functions for the SWE Namespace

These classes represent elements from the OpenGIS(R) Sensor Model Language (SensorML) Implementation Specification that are used to model observation data in responses from a Sensor Observation Service.

SweCompositePhenomenon(id, name, description = as.character(NA), dimension,
                       components, base = NULL)
SwePhenomenon(id, name, description = as.character(NA))
SwePhenomenonProperty(href = as.character(NA), phenomenon = NULL)
SweTextBlock(tokenSeparator, blockSeparator, decimalSeparator, id = as.character(NA))

Arguments

id
The character string to be used for the id attribute (mandatory).

name
The character string to be used for the name element (mandatory).

description
The character string to be used for the description element.

dimension
The dimensions of a composite phenomenon (mandatory).

components
The (sub-) components of a composite phenomenon (mandatory).

base
The (optional) base element for a composite phenomenon.

href
A reference to an (online) object instead of a inline property.

phenomenon
The inline phenomenon of a phenomenon property.

tokenSeparator
The character to be used as the token seperator, often ",".

blockSeparator
The character to be used as the block seperator, often ";".

decimalSeparator
The character to be used as the decimal seperator, often ".".

Objects from the Class

Objects can be created by calls to the constrction functions of the form SweCompositePhenomenon( ...), SwePhenomenonProperty( ...) and so forth.

The following classes are virtual, no objects may be created from them: SwePhenomenonOrNULL, SwePhenomenonPropertyOrNULL.

Slots

dimension:

Object of class "integer", the value of the dimensions attribute of a composite phenomenon.

components:

Object of class "list", the components of a composite phenomenon.

base:

Object of class "SwePhenomenonPropertyOrNULL", the base of a composite phenomenon, may be NULL.

id:

Object of class "character", the value of the id attribute of a (composite) phenomenon.

name:

Object of class "character", the value of the name element of a (composite) phenomenon.

description:

Object of class "character", the value of the description elements of a phenomenon.

href:

Object of class "character", the value of the href attribute of a phenomenon property which references a phenomenon.

phenomenon:

Object of class "SwePhenomenonOrNULL", the actual (inline) phenomenon of a phenomenon property.

tokenSeparator:

Object of class "character", the symbol to be used as the token seperator in a SweTextBlock, e.g. in the case of "," this would result in attribute1,attribute2.

blockSeparator:

Object of class "character", the symbol to be used as the block seperator in a SweTextBlock, e.g. in the case of ";" this would result in attribute1_a,attribute2_a;attribute1_b,attribute2_b.

decimalSeparator:

Object of class "character", the symbol to be used as the decimal seperator in a SweTextBlock, e.g. in the case of "." this would result in attribute1,42.0,attribute3,23.0.

Extends

SweCompositePhenomenon: Class "SwePhenomenon", directly. Class "SwePhenomenonOrNULL", by class "SwePhenomenon", distance 2.

SwePhenomenon: Class "SwePhenomenonOrNULL", directly.

SwePhenomenonProperty: Class "SwePhenomenonPropertyOrNULL", directly.

Methods

show

signature(object = "SweCompositePhenomenon"): ...

show

signature(object = "SwePhenomenon"): ...

show

signature(object = "SwePhenomenonProperty"): ...

show

signature(object = "SweTextBlock"): ...

Details

The ...OrNULL classes are used to model optional slots.

Value

The construction functions return an object of the respective class.

References

See section 9, SWE Common XML Encoding and Examples, of Botts, M., Robin, A. (Eds.), OpenGIS(R) Sensor Model Language (SensorML) Implementation Specification, Open Geospatial Consortium Inc., OGC 07-000

Examples

showClass("SweCompositePhenomenon")

#> Class "SweCompositePhenomenon" [package "sos4R"]
#> 
#> Slots:
#>                                                               
#> Name:                    dimension                  components
#> Class:                     integer                        list
#>                                                               
#> Name:                         base                          id
#> Class: SwePhenomenonPropertyOrNULL                   character
#>                                                               
#> Name:                         name                 description
#> Class:                   character                   character
#> 
#> Extends: 
#> Class "SwePhenomenon", directly
#> Class "SwePhenomenonOrNULL", by class "SwePhenomenon", distance 2

showClass("SwePhenomenon")

#> Class "SwePhenomenon" [package "sos4R"]
#> 
#> Slots:
#>                                           
#> Name:           id        name description
#> Class:   character   character   character
#> 
#> Extends: "SwePhenomenonOrNULL"
#> 
#> Known Subclasses: "SweCompositePhenomenon"

showClass("SwePhenomenonProperty")

#> Class "SwePhenomenonProperty" [package "sos4R"]
#> 
#> Slots:
#>                                               
#> Name:                 href          phenomenon
#> Class:           character SwePhenomenonOrNULL
#> 
#> Extends: "SwePhenomenonPropertyOrNULL"

showClass("SwePhenomenonPropertyOrNULL")

#> Virtual Class "SwePhenomenonPropertyOrNULL" [package "sos4R"]
#> 
#> No Slots, prototype of class "NULL"
#> 
#> Known Subclasses: 
#> Class "NULL", directly
#> Class "SwePhenomenonProperty", directly
#> Class ".NULL", by class "NULL", distance 2, with explicit coerce

showClass("SweTextBlock")

#> Class "SweTextBlock" [package "sos4R"]
#> 
#> Slots:
#>                                                                           
#> Name:    tokenSeparator   blockSeparator decimalSeparator               id
#> Class:        character        character        character        character

if (FALSE) {
sos <- SOS(...)
obj <- parseX(...)
switchSweNamespace(xml2::xml_ns(obj), sos)
}