Bindings and Connecition Methods of OGC Sensor Observation Service

The SOS comes with three possible methods of transfering data, HTTP GET, HTTP POST and SOAP.

Details

The POST binding is described in the official SOS specification and should be the default method.

The GET binding is described by OOTethys in a Best Practice document: https://web.archive.org/web/20120616065001/http://www.oostethys.org/best-practices/best-practices-get. It contains some special encoding for bounding boxes, as the only spatial filter, and time periods, as the only temporal filter.

The SOAP binding is not official with regards to the spec, and also not implemented yet.

The connection method can be changed on creation of a SOS object.

References

Wikipedia page for HTTP request methods: https://en.wikipedia.org/wiki/Hypertext_Transfer_Protocol#Request_methods.

See also

Examples

# HTTP connection methods supported by this sos4R implementation
supported <- SosSupportedBindings()
supported

#> Key-value-pair (GET) Plain old XML (POST) 
#>                "KVP"                "POX" 

if (FALSE) {
sos <- SOS("http://sosurl.org/", binding = "KVP")
}