Class and Construction Function of "SosContents"

SosContents models the sos:Contents section in a service metadata document.

SosContents(observationOfferings)

Arguments

observationOfferings
A list of objects of class SosObservationOffering.

Objects from the Class

Objects can be created by calls to the construction funtion in the form SosContents(...).

The virtual class SosContentsOrNULL is used to model optional slots in classes containing SosContents: No objects may be created from it.

Slots

observationOfferings:

Object of class "list" ~~

xml:

Object of class "XMLAbstractNode" ~~

Extends

Class "OwsContents", directly. Class "SosContentsOrNULL", directly. Class "OwsContentsOrNULL", by class "OwsContents", distance 2.

Methods

show

signature(object = "SosContents"): ...

Details

The SosContents section extends the generic ows:Contents elements. It contains the SosObservationOfferings of a Sensor Observation Service.

References

See section 8.2.3.2, “Contents Section”, of the SOS specification.

See also

Examples

showClass("SosContents")

#> Class "SosContents" [package "sos4R"]
#> 
#> Slots:
#>                                                 
#> Name:  observationOfferings                  xml
#> Class:                 list                  ANY
#> 
#> Extends: 
#> Class "OwsContents", directly
#> Class "SosContentsOrNULL", directly
#> Class "OwsContentsOrNULL", by class "OwsContents", distance 2

showClass("SosContentsOrNULL")

#> Virtual Class "SosContentsOrNULL" [package "sos4R"]
#> 
#> No Slots, prototype of class "NULL"
#> 
#> Known Subclasses: 
#> Class "NULL", directly
#> Class "SosContents", directly
#> Class ".NULL", by class "NULL", distance 2, with explicit coerce