Classes and Construction Functions for "SosFilter_Capabilities" Elements

Additional section in the service metadata document of a Sensor Observation Service, which contains information about the supported filters.

SosFilter_Capabilities(spatial = list(NA_character_), temporal = list(NA_character_),
                       scalar = list(NA_character_), id = list(NA_character_))

Arguments

spatial
A character list of names of available spatial filters.

temporal
A character list of names of available temporal filters.

scalar
A character list of names of available scalar filters.

id
A character list of names of available filters on identifiers.

Objects from the Class

Objects can be created by calls of the form new("SosFilter_Capabilities", ...).

SosFilter_CapabilitiesOrNULL is virtual class: No objects may be created from it.

Slots

spatial:

Object of class "list" with character strings for names of spatial filters.

temporal:

Object of class "list" with character strings for names of temporal filters.

scalar:

Object of class "list" with character strings for names of scalar filters.

id:

Object of class "list" with character strings for names of ID filters.

Extends

Class "SosFilter_CapabilitiesOrNULL", directly.

Methods

show

signature(object = "SosFilter_Capabilities"): ...

Details

The FilterCapabilities section is used to indicate what types of query parameters are supported by the service. These capabilities refer to the parameters of the GetObservation operation which is the only operation that includes OGC filter-like expressions.

References

See section 8.2.3.1, “FilterCapabilities Section”, the SOS specification: Na, A., Priest, M. (Eds.), Sensor Observation Service, Open Geospatial Consortium Inc., OGC 06-009r6, Version: 1.0

See also

Examples

showClass("SosFilter_Capabilities")

#> Class "SosFilter_Capabilities" [package "sos4R"]
#> 
#> Slots:
#>                                           
#> Name:   spatial temporal   scalar       id
#> Class:     list     list     list     list
#> 
#> Extends: "SosFilter_CapabilitiesOrNULL"

showClass("SosFilter_CapabilitiesOrNULL")

#> Virtual Class "SosFilter_CapabilitiesOrNULL" [package "sos4R"]
#> 
#> No Slots, prototype of class "NULL"
#> 
#> Known Subclasses: 
#> Class "NULL", directly
#> Class "SosFilter_Capabilities", directly
#> Class ".NULL", by class "NULL", distance 2, with explicit coerce