Classes and Related Functions for "SosObservationOffering"

SosObservationOfferings collect all metadata about a specific offerign in a Sensor Observation Service.

SosObservationOffering(id, name = as.character(NA), time, procedure, observedProperty,
                       featureOfInterest, responseFormat,
                       intendedApplication = as.character(NA),
                       resultModel = as.character(NA),
                       responseMode = as.character(NA), boundedBy = list())

Arguments

boundedBy
See the corresponding slot description.

featureOfInterest
See the corresponding slot description.

id
See the corresponding slot description.

intendedApplication
See the corresponding slot description.

name
See the corresponding slot description.

observedProperty
See the corresponding slot description.

procedure
See the corresponding slot description.

responseFormat
See the corresponding slot description.

responseMode
See the corresponding slot description.

resultModel
See the corresponding slot description.

time
See the corresponding slot description.

Objects from the Class

Objects can be created by calls to the construction functions of the form SosObservationOffering(...).

Slots

boundedBy:

Object of class "list": A bounding box that contains all features in this offering.

featureOfInterest:

Object of class "list": Features or feature collections that represent the identifiable object(s) on which the sensor systems are making observations.

id:

Object of class "character": Identifier of an offering.

intendedApplication:

Object of class "list": The intended category of use for this offering such as homeland security or natural resource planning

name:

Object of class "character": The name of an offering.

observedProperty:

Object of class "list": The observable/phenomenon that can be requested in this offering.

procedure:

Object of class "list": A reference to one or more procedures, including sensor systems, instruments, simulators, etc, that supply observations in this offering. The DescribeSensor operation can be called to provide a detailed description of each system.

responseFormat:

Object of class "list": MIME type of the data that will be returned as the result of a GetObservation request. This is usually text/xml; subtype="om/0.0.0".

responseMode:

Object of class "list": Indicates what modes of response are supported for this offering. The value of resultTemplate is used to retrieve an observation template that will later be used in calls to GetResult. The other options allow results to appear inline in a resultTag (inline), external to the observation element (out-of-band) or as a MIME attachment (attached).

resultModel:

Object of class "list": Indicates the namespace-qualified name of the result element that will be included in the document returned from a call to GetObservation for this offering, e.g. "om:Observation" or "om:Measurement".

time:

Object of class "GmlTimeGeometricPrimitive": Time period for which observations can be obtained. This supports the advertisement of historical as well as real-time observations.

Methods

sosTime

signature(obj = "SosObservationOffering"): Accessor function for the time slot, or to be more precise: the time period for which this offering provides data.

Details

ObservationOffering provides a mechanism for factoring groups of related observations within a single service instance. A functionally equivalent outcome could be obtained by factoring between different service instances.

Value

The construction functions return an object of the respective class, e.g. SosObservationOffering.

References

See section 6.3, “Observation Offerings”, of the SOS specification: Na, A., Priest, M. (Eds.), Sensor Observation Service, Open Geospatial Consortium Inc., OGC 06-009r6, Version: 1.0

See also

Examples

showClass("SosObservationOffering")

#> Class "SosObservationOffering" [package "sos4R"]
#> 
#> Slots:
#>                                                           
#> Name:                         id                      name
#> Class:                 character                 character
#>                                                           
#> Name:                       time                 procedure
#> Class: GmlTimeGeometricPrimitive                 character
#>                                                           
#> Name:           observedProperty         featureOfInterest
#> Class:                      list                      list
#>                                                           
#> Name:             responseFormat       intendedApplication
#> Class:                 character                 character
#>                                                           
#> Name:                resultModel              responseMode
#> Class:                 character                 character
#>                                 
#> Name:                  boundedBy
#> Class:                      list

showClass("SosObservationOffering_2.0.0")

#> Class "SosObservationOffering_2.0.0" [package "sos4R"]
#> 
#> Slots:
#>                                                             
#> Name:                          id                       name
#> Class:                  character                  character
#>                                                             
#> Name:                  resultTime             phenomenonTime
#> Class:  GmlTimeGeometricPrimitive  GmlTimeGeometricPrimitive
#>                                                             
#> Name:                   procedure         observableProperty
#> Class:                  character                       list
#>                                                             
#> Name:       featureOfInterestType            observationType
#> Class:                       list                       list
#>                                                             
#> Name:                observedArea procedureDescriptionFormat
#> Class:                       list                       list
#>                                  
#> Name:              responseFormat
#> Class:                       list

# explore offerings of an SOS
mySOS <- SOS(url = "http://sensorweb.demo.52north.org/sensorwebtestbed/service/kvp",
             binding = "KVP")
offering1 <- sosOfferings(mySOS)[[1]]

sosId(offering1)

#> [1] "ws2500"

sosName(offering1)

#> [1] NA

sosTime(offering1)

#> $begin
#> [1] "2015-05-18 08:35:00 UTC"
#> 
#> $end
#> [1] "2018-01-06 05:45:00 UTC"
#> 

sosBoundedBy(offering1)

#> $srsName
#> [1] "urn:ogc:def:crs:EPSG::4326"
#> 
#> $lowerCorner
#> [1] "51.934814453125 7.652428150177"
#> 
#> $upperCorner
#> [1] "51.934814453125 7.652428150177"
#>