Functions to Access Supported Features of the Current sos4R Implementation

These functions can be used to access the supported parameters for a range of settings of a SOS connection.

SosSupportedOperations(version = sos100_version)
SosSupportedComparisonOperators()
SosSupportedBindings()
SosSupportedGeometryOperands()
SosSupportedResponseFormats()
SosSupportedResponseModes()
SosSupportedResultModels()
SosSupportedSpatialOperators()
SosSupportedTemporalOperators()
SosSupportedServiceVersions()

Arguments

version
The SOS specification version.

Details

Supported features, like connection methods and supported response modes, are accessible by functions starting with "SosSupported". See the examples section for a complete list of these functions.

It is encouraged to rather use these methods than manually set character values for compatibility with future versions, e.g. SosSupportedBindings()[[1]] instead of directly writing "GET".

Value

A list of supported values for the respective parameter.

See also

Examples

# The supported operations of the specification
SosSupportedOperations()

#> [1] "GetCapabilities"    "DescribeSensor"     "GetObservation"    
#> [4] "GetObservationById"

# HTTP connection methods supported by this sos4R implementation
SosSupportedBindings()

#> Key-value-pair (GET) Plain old XML (POST) 
#>                "KVP"                "POX" 

myBinding <- SosSupportedBindings()[[1]]
myBinding

#> [1] "KVP"

# Formats, modes and models that can be processed by this implementation
SosSupportedResponseFormats()

#> [1] "text/xml;subtype=\"om/1.0.0\""       "text/xml;subtype=\"sensorML/1.0.1\""
#> [3] "text/csv"                           

SosSupportedResultModels()

#> [1] "om:Measurement" "om:Observation"

SosSupportedResponseModes()

#> [1] "inline"

# Operators and operands for filtering in a GetObservation request
SosSupportedTemporalOperators()

#> $TM_After
#> [1] "TM_After"
#> 
#> $TM_Before
#> [1] "TM_Before"
#> 
#> $TM_During
#> [1] "TM_During"
#> 
#> $TM_Equals
#> [1] "TM_Equals"
#> 
#> $`fes:During`
#> [1] "fes:During"
#> 

SosSupportedSpatialOperators()

#> $BBOX
#> [1] "BBOX"
#> 
#> $Contains
#> [1] "Contains"
#> 
#> $Intersects
#> [1] "Intersects"
#> 
#> $Overlaps
#> [1] "Overlaps"
#> 

SosSupportedGeometryOperands()

#> $`gml:Envelope`
#> [1] "gml:Envelope"
#> 
#> $`gml:Polygon`
#> [1] "gml:Polygon"
#> 
#> $`gml:Point`
#> [1] "gml:Point"
#> 
#> $`gml:LineString`
#> [1] "gml:LineString"
#> 

SosSupportedComparisonOperators()

#> named list()