Classes and Construction Methods for Temporal Operator Classes

Classes for temporal operators from OpenGIS(R) Filter Encoding used in filters in GetObservation requests.

TM_After(propertyName = sosDefaultTempOpPropertyName, time)
TM_Before(propertyName = sosDefaultTempOpPropertyName, time)
TM_During(propertyName = sosDefaultTempOpPropertyName, time)
TM_Equals(propertyName = sosDefaultTempOpPropertyName, time)

Arguments

propertyName
The name of the property that is used to wrap the time.

time
A time instant or period to be used as the temporal operand.

Value

An object of the respective class, so TM_After, TM_Before, TM_During or TM_Equals

Objects from the Class

Objects can be created by calls of the form new("TM_After", ...).

Slots

time:

Object of class "GmlTimeInstant"

propertyName:

Object of class "character"

Extends

Class "OgcBinaryTemporalOp", directly. Class "OgcBinaryTemporalOpOrNULL", by class "OgcBinaryTemporalOp", distance 2.

Methods

encodeXML

signature(obj = "TM_After", sos = "SOS"): ...

show

signature(object = "TM_After"): ...

encodeXML

signature(obj = "TM_Before", sos = "SOS"): ...

show

signature(object = "TM_Before"): ...

encodeXML

signature(obj = "TM_During", sos = "SOS"): ...

show

signature(object = "TM_During"): ...

encodeXML

signature(obj = "TM_Equals", sos = "SOS"): ...

show

signature(object = "TM_Equals"): ...

References

Vretanos, Panagiotis A. (Ed.), OpenGIS(R) Filter Encoding Implementation Specification, OGC 04-095, Version: 1.1.0

See the schema file: http://schemas.opengis.net/sos/1.0.0/ogc4sos.xsd.

See also

Examples

showClass("TM_After")

#> Class "TM_After" [package "sos4R"]
#> 
#> Slots:
#>                                     
#> Name:            time   propertyName
#> Class: GmlTimeInstant      character
#> 
#> Extends: 
#> Class "OgcBinaryTemporalOp", directly
#> Class "OgcBinaryTemporalOpOrNULL", by class "OgcBinaryTemporalOp", distance 2

showClass("TM_Before")

#> Class "TM_Before" [package "sos4R"]
#> 
#> Slots:
#>                                     
#> Name:            time   propertyName
#> Class: GmlTimeInstant      character
#> 
#> Extends: 
#> Class "OgcBinaryTemporalOp", directly
#> Class "OgcBinaryTemporalOpOrNULL", by class "OgcBinaryTemporalOp", distance 2

showClass("TM_During")

#> Class "TM_During" [package "sos4R"]
#> 
#> Slots:
#>                                   
#> Name:           time  propertyName
#> Class: GmlTimePeriod     character
#> 
#> Extends: 
#> Class "OgcBinaryTemporalOp", directly
#> Class "OgcBinaryTemporalOpOrNULL", by class "OgcBinaryTemporalOp", distance 2

showClass("TM_Equals")

#> Class "TM_Equals" [package "sos4R"]
#> 
#> Slots:
#>                                                           
#> Name:               propertyName                      time
#> Class:                 character GmlTimeGeometricPrimitive
#> 
#> Extends: 
#> Class "OgcBinaryTemporalOp", directly
#> Class "OgcBinaryTemporalOpOrNULL", by class "OgcBinaryTemporalOp", distance 2

if (FALSE) {
# create times to use for operators
t1 <- sosCreateTimeInstant(sos = weathersos, time = Sys.time())
p1 <- sosCreateTimePeriod(sos = weathersos, begin = as.POSIXct("2010-03-01 12:15"),
                          end = as.POSIXct("2010-03-02 12:15"))

# create temporal operator
afterNow <- TM_After(time = t1)
print(afterNow)
encodeXML(t1, sos)

during <- TM_During(time = p1)
print(during)
}