Methods for Encoding Requests to SOS in SOAP Format

These methods encode objects representing requests to a Sensor Observation Service into a SOAP message format to be used in the SOAP binding (see SosBindings).

Methods

signature(obj = "SosDescribeSensor")

Encode a DescribeSensor operation.

signature(obj = "SosGetObservation")

Encode a GetObservation operation.

signature(obj = "SosGetObservationById")

Encode a GetObservationById operation.

signature(obj = "OwsGetCapabilities")

Encode a GetCapabilities operation.

See also