This function queries a SOS v2.0 for time series data

For given sites and phenomena, a SOS is queried and the data returned as data.frame.

getData(sos, phenomena, sites, spatialBBox = NA, timeInterval = NA_character_)

Arguments

sos
An SOS object from a call to SOS containing the URL of a SOS v2.0.

phenomena
A list, vector or one-column data.frame with characters identifying the relevant phenomena.

sites
A list, vector or one-column data.frame with characters identifying the relevant sites.

spatialBBox
A 2-by-2 matrix with x and y in the rows and min and max in the columns. See Spatial for details. If sites is set, this argument is ignored.

timeInterval
A character defining an ISO-String (ISO 8601), see also xts.

Value

A data.frame containing the data in long form where each row contains a siteID, the time stamp of the observation and the observed measurements for each phenomenon in a separate column. The measurements are provided with units using units.

Examples

if (FALSE) {
mySos <- SOS(url = "https://climate-sos.niwa.co.nz/", binding = "KVP", useDCPs = FALSE, version = "2.0.0")

observationData <- getData(sos = mySos,
                           phenomena = phenomena[18,1],
                           sites = siteList[1:2,1])
}