Function retrieve phenomena of a SOS v2.0

This function queries a SOS v2.0 for all its phenomena

phenomena(sos, includeTemporalBBox = FALSE, listSites = FALSE)

Arguments

sos
An SOS object from a call to SOS containing the URL of a SOS.

includeTemporalBBox
Whether the observation containing timespan of each phenomenon shall also be reported.

listSites
Whether the sites shall be reported at which each phenomenon has been observed.

Value

If only the argument sos is provided, this function returns a character vector with the phenomenon ids.

If timeInfo = TRUE, the result is a data.frame where the character vector of phenomena is combined with a character columns beginTime and endTime. Note that these two columns indicate the very beginning and end of any time series recorded for each phenomenon across all sites and does not imply any continuity.

If sites = TRUE, the result is a data.frame where the character vector of phenomena is combined with a character column siteID. Phenomena that have been observed at several sites will be replicated for each site. If also timeInfo = TRUE, the data.frame will have character columns beginTime and endTime per phenomenon and site. Note that these two columns indicate the very beginning and end of any time series recorded for each phenomenon at each site, but does still not imply to be continuous series.

Examples

## The function is currently defined as
function (sos, includeTemporalBBox = FALSE, listSites = FALSE)
{
    stopifnot(inherits(sos, "SOS_2.0.0"))
    stopifnot(is.logical(timeInfo))
    stopifnot(is.logical(sites))
  }

#> function (sos, includeTemporalBBox = FALSE, listSites = FALSE)
#> {
#>     stopifnot(inherits(sos, "SOS_2.0.0"))
#>     stopifnot(is.logical(timeInfo))
#>     stopifnot(is.logical(sites))
#>   }
#> <environment: 0x55fd06b9b2b0>