This function queries a SOS v2.0 for its sites and returns them as data.frame

Information about the registered sites in a SOS can be queried using this function. This function allows for different levels of detail and returns a vector or data.frame.

siteList(sos, empty = FALSE, timeInterval = NA_character_,
         includePhenomena = FALSE, includeTemporalBBox = FALSE,
         phenomena = list())

Arguments

sos
A SOS object from a call to SOS containing the URL of a SOS.

empty
Whether also empty sites shall be included in the response.

timeInterval
A character defining an ISO-String (ISO 8601), see also xts. Only sites with data within this interval will be retrieved.

includePhenomena
Whether the phenomena shall be listed per location that can be observed.

includeTemporalBBox
Whether the observation containing timespan of each phenomenon shall also be reported.

phenomena
A list, vector or one-column data.frame with characters identifying the relevant phenomena; only sites where these phenomena can be observed are retrieved.

Value

If only the argument sos is provided and the flag empty is set to TRUE or FALSE, this function returns a character vector with the site ids.

If listPhenomena = TRUE, the data.frame contains the column containing the site ids and a character column phenomenon where sites with several phenomena are repeated for each of their phenomenon.

If includeTemporalBBox = TRUE, the data.frame also contains two character columns beginTime and endTime indicating per site id and phenomenon its earliest and latest observation time point.

Examples

## The function is currently defined as
function (sos, empty = FALSE, timeInterval = NA_character_, listPhenomena = FALSE,
    includeTemporalBBox = FALSE, phenomena = list())
{
    stopifnot(inherits(sos, "SOS_2.0.0"))
    stopifnot(is.logical(empty))
    stopifnot(is.character(timeInterval))
    stopifnot(is.logical(listPhenomena))
    stopifnot(is.logical(includeTemporalBBox))
    phenomena <- .validateListOrDfColOfStrings(phenomena, "phenomena")
    if (includeTemporalBBox && !listPhenomena) {
        listPhenomena <- TRUE
        warning("'listPhenomena' has been set to 'TRUE' as this is required for
                 'includeTemporalBBox'.")
    }
  }

#> function (sos, empty = FALSE, timeInterval = NA_character_, listPhenomena = FALSE,
#>     includeTemporalBBox = FALSE, phenomena = list())
#> {
#>     stopifnot(inherits(sos, "SOS_2.0.0"))
#>     stopifnot(is.logical(empty))
#>     stopifnot(is.character(timeInterval))
#>     stopifnot(is.logical(listPhenomena))
#>     stopifnot(is.logical(includeTemporalBBox))
#>     phenomena <- .validateListOrDfColOfStrings(phenomena, "phenomena")
#>     if (includeTemporalBBox && !listPhenomena) {
#>         listPhenomena <- TRUE
#>         warning("'listPhenomena' has been set to 'TRUE' as this is required for
#>                  'includeTemporalBBox'.")
#>     }
#>   }
#> <environment: 0x55d89ed80450>