This function queries a SOS v2.0 for its sites and returns them as SpatialPointsDataFrame

Information about the registered sites in a SOS can be queried using this function. This function allows for different levels of detail and returns a SpatialPointsDataFrame.

sites(sos,
      begin = NA,
      end = NA,
      phenomena = list(),
      empty = FALSE,
      includePhenomena = FALSE,
      includeTemporalBBox = FALSE,
      ...)

Arguments

sos
A SOS object from a call to SOS containing the URL of a SOS.

begin
Object of class POSIXt. Note that creating timestamps with as.POSIXct(..) will use your local timezone.

end
Object of class POSIXt. Note that creating timestamps with as.POSIXct(..) will use your local timezone.

phenomena
A list, vector or one-column data.frame with characters identifying the relevant phenomena.

empty
Whether also empty sites shall be included in the response.

includePhenomena
Whether the phenomena shall also be reported in the result.

includeTemporalBBox
Whether the observation containing timespan of each phenomenon shall also be reported.

...
Named parameters passed on to other wrapper functions.

Value

A SpatialPointsDataFrame containing all sites that match the filter conditions with corresponding metadata defined via the arguments. If empty = TRUE, the data slot contains a logical column 'empty' indicating whether any data has been observed for this site. If listPhenomena = TRUE, the data.frame contains a logical column per phenomenon indicating whether this specific phenomenon has been observed at that site. If includeTemporalBBox = TRUE, the data.frame contains a two column data.frame with beginTime and endTime per phenomenon.

Examples

mySos <- SOS(url = "https://climate-sos.niwa.co.nz/",
             binding = "KVP", useDCPs = FALSE, version = "2.0.0")

sites(sos = mySos)

#>              coordinates siteID
#> 1     (173.926, -35.183)   1056
#> 2   (172.851, -42.53433)  11234
#> 3  (172.9716, -41.09798)  12429
#> 4  (173.9628, -41.49891)  12430
#> 5  (169.3148, -45.20724)  12431
#> 6  (174.9844, -40.90392)  12442
#> 7  (168.3305, -46.41727)  12444
#> 8     (167.275, -45.525)  12482
#> 9  (166.7612, -77.84935)  12740
#> 10 (170.5147, -45.90129)  15752
#> 11    (172.324, -41.805)  16826
#> 12  (175.735, -37.87683)  17030
#> 13 (173.2629, -35.13352)  17067
#> 14 (172.6111, -43.32858)  17244
#> 15     (171.672, -43.47)  17610
#> 16 (-176.475, -43.81687)  17840
#> 17    (170.096, -43.736)  18125
#> 18    (174.867, -41.407)  18234
#> 19 (170.1356, -45.51814)  18437
#> 20  (175.545, -39.19589)  18464
#> 21 (174.0727, -41.64724)  18468
#> 22 (170.1004, -45.12427)  18593
#> 23 (177.5294, -38.28566)   1905
#> 24 (174.8638, -37.20637)   2006
#> 25 (173.0948, -41.31727)  21937
#> 26 (175.3898, -41.25231)  21938
#> 27 (175.6092, -40.38195)  21963
#> 28 (174.7764, -36.96177)  22719
#> 29  (174.305, -39.33546)  23872
#> 30 (171.1916, -42.46022)  23934
#> 31 (172.6077, -43.53074)  24120
#> 32 (170.1343, -43.36548)  24926
#> 33 (177.9218, -38.62747)  24976
#> 34 (173.8532, -35.93145)  25119
#> 35  (169.267, -43.71882)  25506
#> 36 (174.7202, -38.62033)  25726
#> 37 (172.1557, -41.27058)  25777
#> 38 (171.5628, -42.94152)  25821
#> 39 (175.3052, -37.77389)  26117
#> 40 (169.7315, -46.29282)  26163
#> 41 (176.1103, -40.20812)  26958
#> 42 (175.4128, -39.41755)  31621
#> 43 (173.2241, -42.82722)  31832
#> 44 (175.8575, -40.68048)  31851
#> 45 (171.3108, -44.12476)  35704
#> 46 (169.3921, -45.25366)  36592
#> 47 (172.9657, -43.80938)  36593
#> 48     (175.58, -38.517)  37016
#> 49 (167.6825, -45.29644)  37047
#> 50 (177.1629, -38.33334)  37239
#> 51  (168.018, -44.97781)  37382
#> 52 (175.1967, -39.07486)  37850
#> 53 (174.7138, -36.74827)  37852
#> 54 (176.4489, -40.57728)  38057
#> 55 (175.9159, -40.50652)  38224
#> 56 (172.0943, -43.83207)  38866
#> 57 (172.0842, -43.47962)  39063
#> 58 (171.8601, -42.11578)   3925
#> 59 (169.3182, -45.20342)  39564
#> 60  (169.6842, -45.0401)   5535