Constants in sos4R

The package sos4R comes with a set of constant character strings and fixed supported features, for example for names of XML elements, XML Namespace prefixes, or supported formats and models.

Details

Most of these variables should be pretty self-explanatory.

Constants for names of XML elements start with a lowercase character string of the namespace prefix (e.g. "gml"), a unique name of the element (where parts like "type" and special characters may be left out, and other descripte elements may be added for clarity), and end with "Name".

Examples: codegmlEnvelopeName, ogcGeometryOperandLineStringName, ogcTempOpTMEqualsName.

The OwsExceptionsData() function provides access to the fixed exception codes, meanings and respective HTTP codes and messages.

The mime types are used for automatic detection of the best fitting parser.

References

See also Defaults for default parameter settings.

Whiteside A. (Ed.), OGC Web Services Common Specification, Open Geospatial Consortium Inc., OGC 06-121r3, Version: 1.1.0 with Corrigendum 1

Examples

# example constants
sos100NamespacePrefix

#> [1] "sos"

gmlNameName

#> [1] "gml:name"

sweUomName

#> [1] "swe:uom"

# Data frame holding OWS exception code information
OwsExceptionsData()

#>              exceptionCode
#> 1    OperationNotSupported
#> 2    MissingParameterValue
#> 3    InvalidParameterValue
#> 4 VersionNegotiationFailed
#> 5    InvalidUpdateSequence
#> 6       OptionNotSupported
#> 7         NoApplicableCode
#>                                                                                                                                               meaningOfCode
#> 1                                                                                          Request is for an operation that is not supported by this server
#> 2                        Operation request does not include a parameter value, and this server did not declare a default parameter value for that parameter
#> 3                                                                                                     Operation request contains an invalid parameter value
#> 4            List of versions in 'AcceptVersions' parameter value in GetCapabilities operation request did not include any version supported by this server
#> 5 Value of (optional) updateSequence parameter in GetCapabilities operation request is greater than current value of service metadata updateSequence number
#> 6                                                                                             Request is for an option that is not supported by this server
#> 7                                                                     No other exceptionCode specified by this service and server applies to this exception
#>                                locator httpStatusCode           httpMessage
#> 1      Name of operation not supported            501       Not Implemented
#> 2            Name of missing parameter            400           Bad request
#> 3 Name of parameter with invalid value            400           Bad request
#> 4       None, omit 'locator' parameter            400           Bad request
#> 5       None, omit 'locator' parameter            400           Bad request
#> 6   Identifier of option not supported            501       Not implemented
#> 7       None, omit 'locator' parameter  3xx, 4xx, 5xx Internal Server Error

# Get all namespaces
SosAllNamespaces("1.0.0")

#>                                         sos 
#>            "http://www.opengis.net/sos/1.0" 
#>                                         xsi 
#> "http://www.w3.org/2001/XMLSchema-instance" 
#>                                          om 
#>             "http://www.opengis.net/om/1.0" 
#>                                         ows 
#>            "http://www.opengis.net/ows/1.1" 
#>                                         ogc 
#>                "http://www.opengis.net/ogc" 
#>                                         sml 
#>     "http://www.opengis.net/sensorML/1.0.1" 
#>                                         swe 
#>          "http://www.opengis.net/swe/1.0.1" 
#>                                       xlink 
#>              "http://www.w3.org/1999/xlink" 
#>                                         gml 
#>                "http://www.opengis.net/gml" 
#>                                          sa 
#>       "http://www.opengis.net/sampling/1.0"