Classes and Construction Functions for the OGC Namespace

These classes represent elements from the OpenGIS(R) Filter Encoding Implementation Specification that are used in requests to Sensor Observation Services.

OgcBBOX(propertyName = sosDefaultSpatialOpPropertyName, envelope)
OgcContains(propertyName = sosDefaultSpatialOpPropertyName, geometry = NULL,
            envelope = NULL)
OgcIntersects(propertyName = sosDefaultSpatialOpPropertyName, geometry = NULL,
              envelope = NULL)
OgcOverlaps(propertyName = sosDefaultSpatialOpPropertyName, geometry = NULL,
            envelope = NULL)

Arguments

propertyName
The value for the propertyName attribute.

geometry
The geometry to be used in a spatial filter.

envelope
The geometry to be used in a sptial filter.

Objects from the Class

Objects can be created by calls to the respective construction functions of the form OgcBBOX( ...), OgcContains(...), or OgcIntersects.

The following classes are virtual, no objects may be created from them: OgcBinaryTemporalOp, OgcBinaryTemporalOpOrNULL, OgcComparisonOps, codeOgcComparisonOpsOrNULL, OgcSpatialOps, OgcSpatialOpsOrNULL.

Slots

propertyName:

Object of class "character", the value of the propertyName attribute.

geometry:

Object of class "GmlGeometry", a geometry contained in a spatial filter.

envelope:

Object of class "GmlEnvelope", an envelope contained in a spatial filter.

time:

Object of class "GmlTimeGeometricPrimitive", a time element contained in a temporal filter.

Extends

OgcBBOX, OgcBinarySpatialOp: Class "OgcSpatialOps", directly. Class "OgcSpatialOpsOrNULL", by class "OgcSpatialOps", distance 2.

OgcBinaryTemporalOp: Class "OgcBinaryTemporalOpOrNULL", directly.

OgcContains, OgcIntersects, OgcOverlaps: Class "OgcBinarySpatialOp", directly. Class "OgcSpatialOps", by class "OgcBinarySpatialOp", distance 2. Class "OgcSpatialOpsOrNULL", by class "OgcBinarySpatialOp", distance 3.

OgcSpatialOps: Class "OgcSpatialOpsOrNULL", directly.

Methods

encodeKVP

signature(obj = "OgcBinaryTemporalOp", sos = "SOS"): Encode the given operation in key-value-pair style, see encodeKVP

encodeXML

signature(obj = "OgcBBOX", sos = "SOS"): Encode the given operation in XML, see encodeXML

encodeXML

signature(obj = "OgcComparisonOps", sos = "SOS"): Encode the given operation in XML, see encodeXML

encodeXML

signature(obj = "OgcContains", sos = "SOS"): Encode the given operation in XML, see encodeXML

encodeXML

signature(obj = "OgcIntersects", sos = "SOS"): Encode the given operation in XML, see encodeXML

encodeXML

signature(obj = "OgcOverlaps", sos = "SOS"): Encode the given operation in XML, see encodeXML

% \item{show}{\code{signature(object = "OgcBBOX")}: Show the given object in a human readable fashion.} % \item{show}{\code{signature(object = "OgcContains")}: Show the given object in a human readable fashion.} % \item{show}{\code{signature(object = "OgcOverlaps")}: Show the given object in a human readable fashion.}

Value

The value of the construction functions is an object of the respective class.

Details

These comprise spatial and temporal operations and operators which can be encoded in different ways.

The ...OrNULL classes are used to model optional slots.

References

Vretanos, Panagiotis A. (Ed.), OpenGIS(R) Filter Encoding Implementation Specification, OGC 04-095, Version: 1.1.0

Schemas: http://schemas.opengis.net/filter/1.1.0/

Note

This implementation of the Filter Encoding Specification is not complete.

Warning

The encoding functions of these classes are not completely implemented yet.

Examples

showClass("OgcBBOX")

#> Class "OgcBBOX" [package "sos4R"]
#> 
#> Slots:
#>                                 
#> Name:  propertyName     envelope
#> Class:    character  GmlEnvelope
#> 
#> Extends: 
#> Class "OgcSpatialOps", directly
#> Class "OgcSpatialOpsOrNULL", by class "OgcSpatialOps", distance 2

showClass("OgcBinarySpatialOp")

#> Virtual Class "OgcBinarySpatialOp" [package "sos4R"]
#> 
#> Slots:
#>                                              
#> Name:  propertyName     geometry     envelope
#> Class:    character  GmlGeometry  GmlEnvelope
#> 
#> Extends: 
#> Class "OgcSpatialOps", directly
#> Class "OgcSpatialOpsOrNULL", by class "OgcSpatialOps", distance 2
#> 
#> Known Subclasses: "OgcContains", "OgcIntersects", "OgcOverlaps"

showClass("OgcBinaryTemporalOp")

#> Virtual Class "OgcBinaryTemporalOp" [package "sos4R"]
#> 
#> Slots:
#>                                                           
#> Name:               propertyName                      time
#> Class:                 character GmlTimeGeometricPrimitive
#> 
#> Extends: "OgcBinaryTemporalOpOrNULL"
#> 
#> Known Subclasses: "TM_After", "TM_Before", "TM_During", "TM_Equals"

showClass("OgcBinaryTemporalOpOrNULL")

#> Virtual Class "OgcBinaryTemporalOpOrNULL" [package "sos4R"]
#> 
#> No Slots, prototype of class "NULL"
#> 
#> Known Subclasses: 
#> Class "NULL", directly
#> Class "OgcBinaryTemporalOp", directly
#> Class ".NULL", by class "NULL", distance 2, with explicit coerce

showClass("OgcComparisonOps")

#> Virtual Class "OgcComparisonOps" [package "sos4R"]
#> 
#> No Slots, prototype of class "NULL"
#> 
#> Extends: "OgcComparisonOpsOrXMLOrNULL"

showClass("OgcContains")

#> Class "OgcContains" [package "sos4R"]
#> 
#> Slots:
#>                                              
#> Name:  propertyName     geometry     envelope
#> Class:    character  GmlGeometry  GmlEnvelope
#> 
#> Extends: 
#> Class "OgcBinarySpatialOp", directly
#> Class "OgcSpatialOps", by class "OgcBinarySpatialOp", distance 2
#> Class "OgcSpatialOpsOrNULL", by class "OgcBinarySpatialOp", distance 3

showClass("OgcOverlaps")

#> Class "OgcOverlaps" [package "sos4R"]
#> 
#> Slots:
#>                                              
#> Name:  propertyName     geometry     envelope
#> Class:    character  GmlGeometry  GmlEnvelope
#> 
#> Extends: 
#> Class "OgcBinarySpatialOp", directly
#> Class "OgcSpatialOps", by class "OgcBinarySpatialOp", distance 2
#> Class "OgcSpatialOpsOrNULL", by class "OgcBinarySpatialOp", distance 3

showClass("OgcSpatialOps")

#> Virtual Class "OgcSpatialOps" [package "sos4R"]
#> 
#> No Slots, prototype of class "S4"
#> 
#> Extends: "OgcSpatialOpsOrNULL"
#> 
#> Known Subclasses: 
#> Class "OgcBBOX", directly
#> Class "OgcBinarySpatialOp", directly
#> Class "OgcContains", by class "OgcBinarySpatialOp", distance 2
#> Class "OgcIntersects", by class "OgcBinarySpatialOp", distance 2
#> Class "OgcOverlaps", by class "OgcBinarySpatialOp", distance 2

showClass("OgcSpatialOpsOrNULL")

#> Virtual Class "OgcSpatialOpsOrNULL" [package "sos4R"]
#> 
#> No Slots, prototype of class "NULL"
#> 
#> Known Subclasses: 
#> Class "NULL", directly
#> Class "OgcSpatialOps", directly
#> Class ".NULL", by class "NULL", distance 2, with explicit coerce
#> Class "OgcContains", by class "OgcBinarySpatialOp", distance 3
#> Class "OgcIntersects", by class "OgcBinarySpatialOp", distance 3
#> Class "OgcOverlaps", by class "OgcBinarySpatialOp", distance 3