Classes for om:Observation Elements

Classes and construction functions for objects from the OGC Observations and Measurements specification.

OmObservation(samplingTime, procedure, observedProperty, featureOfInterest,
              result, metadata = NA, resultTime = NULL, resultQuality = NA,
              parameter = NA)
OmObservationProperty(href = as.character(NA), obs = NULL)

Arguments

samplingTime
See slot description.

procedure
See slot description.

observedProperty
See slot description.

featureOfInterest
See slot description.

result
See slot description.

metadata
See slot description.

resultTime
See slot description.

resultQuality
See slot description.

parameter
See slot description.

href
See slot description.

obs
See slot description.

Objects from the Class

Objects can be created by calls to the construction functions of the form OmObservation(...) and OmObservationProperty(...).

The following classes are virtual, no objects may be created from them: OmObservationOrNULL-class.

Slots

samplingTime:

Object of class "GmlTimeObjectOrNULL": The time that the result applies to the feature-of-interest. This is the time usually required for geospatial analysis of the result.

procedure:

Object of class "ANY": The description of a process used to generate the result. It must be suitable for the observed property.

observedProperty:

Object of class "SwePhenomenonProperty": Identifies or describes the phenomenon for which the observation result provides an estimate of its value. It must be a property associated with the type of the feature of interest.

featureOfInterest:

Object of class "GmlFeature": A feature of any type (ISO 19109, ISO 19101), which is a representation of the observation target, being the real-world object regarding which the observation is made.

result:

Object of class "ANY": Contains the value generated by the procedure. The type of the observation result must be consistent with the observed property, and the scale or scope for the value must be consistent with the quantity or category type.

metadata:

Object of class "ANY": Observation metadata.

resultTime:

Object of class "GmlTimeObjectOrNULL": The time when the procedure associated with the observation act was applied. For some observations these are identical, in which case the resultTime may be omitted. However, there are important cases where they differ.

resultQuality:

Object of class "ANY": Event specific quality of a result.

parameter:

Object of class "ANY": A general event-specific parameter. This will typically be used to record environmental parameters, or event-specific sampling parameters that are not tightly bound to either the feature-of-interest or the procedure.

href:

Object of class "character": Referenced observation in a OmObservationProperty.

Extends

OmObservation Class "OmObservationOrNULL", directly.

Methods

sosResult

signature(obj = "OmObservation"): Accessor function for the result slot of an observation or measurement.

Details

The class OmObservationProperty can be used to reference to an (online) observation.

The ...OrNULL classes are used to model optional slots.

Value

The construction functions return an object of the respective class.

References

Cox, S. (Ed.), Observations and Measurements - Part 1 - Observation schema, Open Geospatial Consortium Inc., OGC 07-022r1, Version: 1.0

See also

Examples

showClass("OmObservation")

#> Class "OmObservation" [package "sos4R"]
#> 
#> Slots:
#>                                            
#> Name:                          samplingTime
#> Class:                  GmlTimeObjectOrNULL
#>                                            
#> Name:                             procedure
#> Class:                                  ANY
#>                                            
#> Name:                      observedProperty
#> Class:          SwePhenomenonPropertyOrNULL
#>                                            
#> Name:                     featureOfInterest
#> Class: GmlFeatureOrGmlFeaturePropertyOrNULL
#>                                            
#> Name:                                result
#> Class:                                  ANY
#>                                            
#> Name:                              metadata
#> Class:                                  ANY
#>                                            
#> Name:                            resultTime
#> Class:                  GmlTimeObjectOrNULL
#>                                            
#> Name:                         resultQuality
#> Class:                                  ANY
#>                                            
#> Name:                             parameter
#> Class:                                  ANY
#> 
#> Extends: "OmObservationOrNULL"
#> 
#> Known Subclasses: "OmMeasurement"

showClass("OmObservationProperty")

#> Class "OmObservationProperty" [package "sos4R"]
#> 
#> Slots:
#>                                               
#> Name:                 href                 obs
#> Class:           character OmObservationOrNULL

showClass("OmObservationOrNULL")

#> Virtual Class "OmObservationOrNULL" [package "sos4R"]
#> 
#> No Slots, prototype of class "NULL"
#> 
#> Known Subclasses: 
#> Class "NULL", directly
#> Class "OmObservation", directly
#> Class ".NULL", by class "NULL", distance 2, with explicit coerce

OmObservationProperty(href = "http://link.to/myObservation")

#> Object of class OmObservationProperty; href: http://link.to/myObservation; observation:  

# get result from an observation
if (FALSE) {
result <- observation@result

# the accessor method also works with lists of observations
result <- sosResult(observation)
resultList <- sosResult(observationList)
}