Class and Construction Function for "SosCapabilities"

The service metadata document of a Sensor Observation Service.

SosCapabilities(version, updateSequence = NA, owsVersion = "1.1.0",
                identification = NULL, provider = NULL, operations = NULL,
                filterCapabilities = NULL, contents = NULL)

Arguments

version
The version of the service.

updateSequence
Service metadata document version, value is "increased" whenever any change is made in complete service metadata document. This can be used to request a certain version of a metadata document. Parameter is found in both request and reponse, but may not be supported by a service.

owsVersion
The used OWS specification version.

identification
The identification section of a capabilities document, object of class OwsServiceIdentification.

provider
The provider section of a capabilities document, object of class OwsServiceProvider.

operations
A list of objects of class OwsOperation in a OperationsMetadata object. The provider section of a capabilities document.

filterCapabilities
An object of class SosFilter_Capabilities containing the filter capabilities of a service.

contents
The provider section of a capabilities document, object of class SosContents.

Objects from the Class

Objects can be created by calls to the construction function of the form SosCapabilities(...) including the parameter "owsVersion" for the respective version of the service metadata document.

Slots

filterCapabilities:

Object of class "SosFilter_CapabilitiesOrNULL" ~~

identification:

Object of class "OwsServiceIdentificationOrNULL" ~~

provider:

Object of class "OwsServiceProviderOrNULL" ~~

operations:

Object of class "OwsOperationsMetadataOrNULL" ~~

contents:

Object of class "OwsContentsOrNULL" ~~

version:

Object of class "character" ~~

updateSequence:

Object of class "character" ~~

owsVersion:

Object of class "character" ~~

Extends

Class "OwsCapabilities_1.1.0", directly. Class "OwsCapabilities", by class "OwsCapabilities_1.1.0", distance 2.

Methods

No methods defined with class "SosCapabilities_1.0.0" in the signature.

Details

This document provides clients with service metadata about a specific service instance, including metadata about the tightly-coupled data served.

The portions of the GetCapabilities response document that are defined in the OWS Common specification are not modified for SOS. The sections of the response that are specific for the SOS are the Filter_Capabilities and the Contents section.

References

Section 8.2.3 of the SOS specification: Na, A., Priest, M. (Eds.), Sensor Observation Service, Open Geospatial Consortium Inc., OGC 06-009r6, Version: 1.0

See also

Examples

showClass("SosCapabilities_1.0.0")

#> Class "SosCapabilities_1.0.0" [package "sos4R"]
#> 
#> Slots:
#>                                                                     
#> Name:              filterCapabilities                 identification
#> Class:   SosFilter_CapabilitiesOrNULL OwsServiceIdentificationOrNULL
#>                                                                     
#> Name:                        provider                     operations
#> Class:       OwsServiceProviderOrNULL    OwsOperationsMetadataOrNULL
#>                                                                     
#> Name:                        contents                        version
#> Class:              OwsContentsOrNULL                      character
#>                                                                     
#> Name:                  updateSequence                     owsVersion
#> Class:                      character                      character
#> 
#> Extends: 
#> Class "OwsCapabilities_1.1.0", directly
#> Class "OwsCapabilities", by class "OwsCapabilities_1.1.0", distance 2

showClass("SosCapabilities_2.0.0")

#> Class "SosCapabilities_2.0.0" [package "sos4R"]
#> 
#> Slots:
#>                                                                     
#> Name:              filterCapabilities                 identification
#> Class:   SosFilter_CapabilitiesOrNULL OwsServiceIdentificationOrNULL
#>                                                                     
#> Name:                        provider                     operations
#> Class:       OwsServiceProviderOrNULL    OwsOperationsMetadataOrNULL
#>                                                                     
#> Name:                        contents                        version
#> Class:              OwsContentsOrNULL                      character
#>                                                                     
#> Name:                  updateSequence                     owsVersion
#> Class:                      character                      character
#> 
#> Extends: 
#> Class "OwsCapabilities_1.1.0", directly
#> Class "OwsCapabilities", by class "OwsCapabilities_1.1.0", distance 2