Classes and Construction Functions for sos:eventTime elements.

Temporal query parameters for GetObservation requests.

SosEventTime(temporalOps)
SosEventTimeLatest()

Arguments

temporalOps
An object of class OgcBinaryTemporalOp-class to be wrapped by the sos:eventTime element.

Objects from the Classes

Objects can be created by calls to the construction functions of the form SosEventTime(...) or SosEventTimeLatest(...).

Slots

temporalOps:

Object of class "OgcBinaryTemporalOp" for SosEventTime, the temporal operand to be inserted into the event time, or an object of class "character" for SosEventTimeLatest.

Methods

encodeKVP

signature(obj = "SosEventTime", sos = "SOS"): Encode the given object as a key-value pair.

encodeXML

signature(obj = "SosEventTime", sos = "SOS") or signature(obj = "SosEventTimeLatest", sos = "SOS"): Encode the given object as XML.

Details

Specifies the time period(s) for which observations are requested. This allows a client to request observations from a specific instant, multiple instances or periods of time in the past, present and future. The supported range is listed in the selected offering capabilities. The objects of these classes are used in the GetObservation (paramter in GetObservation).

The request for the latest observation is not standard conform. The distinction between the classes happens eventually during encoding, where both classes have different encoding functions.

A typical example in a POST request: <eventTime> <ogc:TM_During> <ogc:PropertyName>om:samplingTime</ogc:PropertyName> <gml:TimePeriod> <gml:beginPosition>2006-11-05T17:18:58.000-06:00</gml:beginPosition> <gml:endPosition>2006-11-05T21:18:59.000-06:00</gml:endPosition> </gml:TimePeriod> </ogc:TM_During> </eventTime>

In GET binding (SosBindings) the eventTime is simply omitted for getting the latest observation.

It is recommended to use the creation functions as shown in the examples.

References

See SOS specification, Table 4: “Parameters of GetObservation Request”.

See also

Examples

showClass("SosEventTime")

#> Class "SosEventTime" [package "sos4R"]
#> 
#> Slots:
#>                           
#> Name:          temporalOps
#> Class: OgcBinaryTemporalOp

showClass("SosEventTimeLatest")

#> Class "SosEventTimeLatest" [package "sos4R"]
#> 
#> Slots:
#>                   
#> Name:  temporalOps
#> Class:   character

# NOT RUN {
# create SosEventTime for all times after the given time stamp
timePos <- GmlTimePosition(as.POSIXct("2010-01-01 12:00"))
tOps <- TM_After(time = GmlTimeInstant(timePosition = timePos))
time1 <- SosEventTime(tOps)

# encode it as XML and KVP
encodeXML(time1)
encodeKVP(time1)

time2 <- SosEventTimeLatest()
encodeXML(time2)
# }