Function retrieve phenomena of a SOS v2.0

This function queries a SOS v2.0 for all its phenomena

phenomena(sos, includeTemporalBBox = FALSE, includeSiteId = FALSE, ...)

Arguments

sos
An SOS object from a call to SOS containing the URL of a SOS.

includeTemporalBBox
Whether the observation containing timespan of each phenomenon shall also be reported.

includeSiteId
Whether the sites shall be reported at which each phenomenon has been observed.

...
Named arguments forwarded to other wrapper functions internally.

Value

If only the argument sos is provided, this function returns a character vector with the phenomenon ids.

If timeInfo = TRUE, the result is a data.frame where the character vector of phenomena is combined with a character columns beginTime and endTime. Note that these two columns indicate the very beginning and end of any time series recorded for each phenomenon across all sites and does not imply any continuity.

If sites = TRUE, the result is a data.frame where the character vector of phenomena is combined with a character column siteID. Phenomena that have been observed at several sites will be replicated for each site. If also timeInfo = TRUE, the data.frame will have character columns beginTime and endTime per phenomenon and site. Note that these two columns indicate the very beginning and end of any time series recorded for each phenomenon at each site, but does still not imply to be continuous series.

Examples

mySos <- SOS(url = "https://climate-sos.niwa.co.nz/",
             binding = "KVP", useDCPs = FALSE, version = "2.0.0")

phenomena(sos = mySos)

#>                                                                                              phenomenon
#> 1                         MTHLY_STATS: DAYS OF DEFICIT (WBal AWC=150mm) (MTHLY: DAYS OF DEFICIT (WBAL))
#> 2                                               MTHLY_STATS: DAYS OF OCCURRENCE (FOG) (MTHLY: FOG DAYS)
#> 3                                             MTHLY_STATS: DAYS OF OCCURRENCE (GALE) (MTHLY: GALE DAYS)
#> 4                             MTHLY_STATS: DAYS OF OCCURRENCE (GROUND FROST) (MTHLY: GROUND FROST DAYS)
#> 5                           MTHLY_STATS: DAYS OF OCCURRENCE (GUSTS over 23 knots) (MTHLY: GUST DAYS 24)
#> 6                           MTHLY_STATS: DAYS OF OCCURRENCE (GUSTS over 32 knots) (MTHLY: GUST DAYS 33)
#> 7                           MTHLY_STATS: DAYS OF OCCURRENCE (GUSTS over 50 knots) (MTHLY: GUST DAYS 51)
#> 8                                             MTHLY_STATS: DAYS OF OCCURRENCE (HAIL) (MTHLY: HAIL DAYS)
#> 9                                   MTHLY_STATS: DAYS OF OCCURRENCE (LIGHTNING) (MTHLY: LIGHTNING DAYS)
#> 10                            MTHLY_STATS: DAYS OF OCCURRENCE (SCREEN FROST) (MTHLY: SCREEN FROST DAYS)
#> 11                                MTHLY_STATS: DAYS OF OCCURRENCE (SNOW LYING) (MTHLY: SNOW LYING DAYS)
#> 12                                            MTHLY_STATS: DAYS OF OCCURRENCE (SNOW) (MTHLY: SNOW DAYS)
#> 13                                      MTHLY_STATS: DAYS OF OCCURRENCE (THUNDER) (MTHLY: THUNDER DAYS)
#> 14                          MTHLY_STATS: DAYS OF RUNOFF (WBal AWC=150mm) (MTHLY: DAYS OF RUNOFF (WBAL))
#> 15         MTHLY_STATS: DAYS WITH MAXIMUM AIR TEMPERATURE OVER 25.0 DEG C. (MTHLY: DAYS OVER 25 DEG C.)
#> 16                            MTHLY_STATS: EXTREME DAILY MAXIMUM UVI (MTHLY: EXTREME DAILY MAXIMUM UVI)
#> 17                               MTHLY_STATS: EXTREME GRASS-MINIMUM TEMPERATURE (MTHLY: EXTR GRASS-MIN)
#> 18                                      MTHLY_STATS: EXTREME MAXIMUM TEMPERATURE (MTHLY: EXTR MAX TEMP)
#> 19                            MTHLY_STATS: EXTREME MAXIMUM TOTAL OZONE (MTHLY: EXTREME MAX TOTAL OZONE)
#> 20                                      MTHLY_STATS: EXTREME MINIMUM TEMPERATURE (MTHLY: EXTR MIN TEMP)
#> 21                            MTHLY_STATS: EXTREME MINIMUM TOTAL OZONE (MTHLY: EXTREME MIN TOTAL OZONE)
#> 22                               MTHLY_STATS: HIGHEST DAILY MEAN TEMPERATURE (MTHLY: HIGHEST MEAN TEMP)
#> 23                                        MTHLY_STATS: HIGHEST MINIMUM TEMPERATURE (MTHLY: HIGHEST MIN)
#> 24                              MTHLY_STATS: LOWEST DAILY MAXIMUM TEMPERATURE (MTHLY: LOWEST DAILY MAX)
#> 25                                 MTHLY_STATS: LOWEST DAILY MEAN TEMPERATURE (MTHLY: LOWEST MEAN TEMP)
#> 26                                          MTHLY_STATS: MAXIMUM 1-DAY RAINFALL (MTHLY: MAX 1-DAY RAIN)
#> 27                                          MTHLY_STATS: MAXIMUM 1-hour rainfall (MTHLY: MAX 1-HR RAIN)
#> 28                                     MTHLY_STATS: MAXIMUM 10-minute rainfall (MTHLY: MAX 10-MIN RAIN)
#> 29                                        MTHLY_STATS: MAXIMUM 12-hour rainfall (MTHLY: MAX 12-HR RAIN)
#> 30                                          MTHLY_STATS: MAXIMUM 2-hour rainfall (MTHLY: MAX 2-HR RAIN)
#> 31                                     MTHLY_STATS: MAXIMUM 20-minute rainfall (MTHLY: MAX 20-MIN RAIN)
#> 32                                        MTHLY_STATS: MAXIMUM 24-hour rainfall (MTHLY: MAX 24-HR RAIN)
#> 33                                     MTHLY_STATS: MAXIMUM 30-minute rainfall (MTHLY: MAX 30-MIN RAIN)
#> 34                                        MTHLY_STATS: MAXIMUM 48-hour rainfall (MTHLY: MAX 48-HR RAIN)
#> 35                                          MTHLY_STATS: MAXIMUM 6-hour rainfall (MTHLY: MAX 6-HR RAIN)
#> 36                                        MTHLY_STATS: MAXIMUM 72-hour rainfall (MTHLY: MAX 72-HR RAIN)
#> 37                             MTHLY_STATS: MAXIMUM DAILY WIND RUN (km) (MTHLY: EXTR DLY WIND RUN (km))
#> 38           MTHLY_STATS: MAXIMUM DEFICIT OF SOIL MOISTURE (WBal AWC=150mm) (MTHLY: MAX DEFICIT (WBAL))
#> 39                           MTHLY_STATS: MAXIMUM GUST DIRECTION (for gust speed) (MTHLY: MAX GUST DIR)
#> 40                       MTHLY_STATS: MAXIMUM GUST SPEED (see 46 for direction) (MTHLY: MAX GUST SPEED)
#> 41                MTHLY_STATS: MAXIMUM SOIL MOISTURE Percent at 20cm depth (MTHLY: MAX SOIL MOISTURE %)
#> 42                                         MTHLY_STATS: MEAN 9AM RELATIVE HUMIDITY (MTHLY: MEAN 9AM RH)
#> 43                                             MTHLY_STATS: MEAN 9AM TEMPERATURE (MTHLY: MEAN 9AM TEMP)
#> 44                             MTHLY_STATS: MEAN AIR TEMPERATURE;   0.5* (MAX + MIN) (MTHLY: MEAN TEMP)
#> 45                                            MTHLY_STATS: MEAN CLOUD AMOUNT (MTHLY: MEAN CLOUD AMOUNT)
#> 46                      MTHLY_STATS: MEAN DAILY ACCUMULATED UV RADIATION (MTHLY: MEAN DAILY ACC UV RAD)
#> 47                        MTHLY_STATS: MEAN DAILY GRASS-MIN from dly Grass-mins (MTHLY: MEAN GRASS-MIN)
#> 48                                  MTHLY_STATS: MEAN DAILY MAXIMUM UVI (MTHLY: MEAN DAILY MAXIMUM UVI)
#> 49                              MTHLY_STATS: MEAN DAILY RADIATION (diFfuse) (MTHLY: MEAN DAILY RAD (F))
#> 50                              MTHLY_STATS: MEAN DAILY RADIATION (diRect). (MTHLY: MEAN DAILY RAD (R))
#> 51                                     MTHLY_STATS: MEAN DAILY RADIATION (Global) (MTHLY: MEAN RAD (G))
#> 52              MTHLY_STATS: MEAN DAILY WIND RUN (km) with NO direction (MTHLY: MEAN DLY WIND RUN (km))
#> 53             MTHLY_STATS: MEAN DEFICIT OF SOIL MOISTURE (WBal AWC=150mm) (MTHLY: MEAN DEFICIT (WBAL))
#> 54                         MTHLY_STATS: MEAN EARTH TEMPERATURE at 100cm depth (MTHLY: MEAN EARTH 100cm)
#> 55                           MTHLY_STATS: MEAN EARTH TEMPERATURE at 10cm depth (MTHLY: MEAN EARTH 10cm)
#> 56                           MTHLY_STATS: MEAN EARTH TEMPERATURE at 20cm depth (MTHLY: MEAN EARTH 20cm)
#> 57                           MTHLY_STATS: MEAN EARTH TEMPERATURE at 30cm depth (MTHLY: MEAN EARTH 30cm)
#> 58                           MTHLY_STATS: MEAN EARTH TEMPERATURE at 50cm depth (MTHLY: MEAN EARTH 50cm)
#> 59                             MTHLY_STATS: MEAN EARTH TEMPERATURE at 5cm depth (MTHLY: MEAN EARTH 5cm)
#> 60                         MTHLY_STATS: MEAN MAXIMUM TEMPERATURE from daily Maxs (MTHLY: MEAN MAX TEMP)
#> 61                         MTHLY_STATS: MEAN MINIMUM TEMPERATURE from daily Mins (MTHLY: MEAN MIN TEMP)
#> 62                                 MTHLY_STATS: MEAN PRESSURE at 0900 local time (MTHLY: MEAN PRESSURE)
#> 63                  MTHLY_STATS: MEAN SOIL MOISTURE Percent at 20cm depth (MTHLY: MEAN SOIL MOISTURE %)
#> 64                                              MTHLY_STATS: MEAN TOTAL OZONE (MTHLY: MEAN TOTAL OZONE)
#> 65                                                   MTHLY_STATS: MEAN VAPOUR PRESSURE (MTHLY: MEAN VP)
#> 66                                 MTHLY_STATS: MEAN WIND SPEED (NO direction) (MTHLY: MEAN WIND SPEED)
#> 67                MTHLY_STATS: MINIMUM SOIL MOISTURE Percent at 20cm depth (MTHLY: MIN SOIL MOISTURE %)
#> 68                                        MTHLY_STATS: RAIN DAYS with 0.1 mm or more (MTHLY: RAIN DAYS)
#> 69                              MTHLY_STATS: STD-DEV OF DAILY MEAN TEMPERATURE (MTHLY: SD OF MEAN TEMP)
#> 70                 MTHLY_STATS: TOTAL COOLING DEGREE DAYS for 18 deg C. (MTHLY: COOLING DEG DAYS (18C))
#> 71                            MTHLY_STATS: TOTAL DEFICIT (WBal AWC=150mm) (MTHLY: TOTAL DEFICIT (WBAL))
#> 72                              MTHLY_STATS: TOTAL EVAPORATION from Raised Pan (MTHLY: TOTAL EVAP (RP))
#> 73                 MTHLY_STATS: TOTAL GROWING DEGREE DAYS for 10 deg C. (MTHLY: GROWING DEG DAYS (10C))
#> 74                   MTHLY_STATS: TOTAL GROWING DEGREE DAYS for 5 deg C. (MTHLY: GROWING DEG DAYS (5C))
#> 75                 MTHLY_STATS: TOTAL HEATING DEGREE DAYS for 18 deg C. (MTHLY: HEATING DEG DAYS (18C))
#> 76                      MTHLY_STATS: TOTAL PENMAN OPEN WATER EVAPORATION (MTHLY: TOTAL OPEN-WATER EVAP)
#> 77                     MTHLY_STATS: TOTAL PENMAN POTENTIAL EVAPO-TRANSPIRATION (MTHLY: TOTAL PENMAN PE)
#> 78 MTHLY_STATS: TOTAL PRIESTLEY-TAYLOR POTENTIAL EVAPO-TRANSPIRATION (MTHLY: TOTAL PRIESTLEY-TAYLOR PE)
#> 79                                                      MTHLY_STATS: TOTAL RAINFALL (MTHLY: TOTAL RAIN)
#> 80                              MTHLY_STATS: TOTAL RUNOFF (WBal AWC=150mm) (MTHLY: TOTAL RUNOFF (WBAL))
#> 81                                                       MTHLY_STATS: TOTAL SUNSHINE (MTHLY: TOTAL SUN)
#> 82                                   MTHLY_STATS: WET DAYS with rainfall 1 mm or more (MTHLY: WET DAYS)
#> 83                                          MTHLY_STATS: WINTER CHILL HOURS (MTHLY: WINTER CHILL HOURS)
#> 84                                          MTHLY_STATS: WINTER CHILL UNITS (MTHLY: WINTER CHILL UNITS)